New changes

+++ src/UI/DeletePopup.java
@@ -1,5 +1,7 @@
 package UI;
 
+import java.awt.Color;
+import java.awt.Font;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
@@ -8,13 +10,16 @@ import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 import javax.swing.JTextField;
+import javax.swing.event.DocumentEvent;
+import javax.swing.event.DocumentListener;
 
 import database.DbRepository;
 
 public class DeletePopup {
+	private Font font = new Font("", Font.PLAIN, 10);
 	private JButton buttonDelete, buttonCancel;
 	private JTextField textFieldId;
-	private JLabel labelId;
+	private JLabel labelId, labelIdShowError;
 	private JFrame frame;
 
 	public DeletePopup() {
@@ -23,6 +28,11 @@ public class DeletePopup {
 
 	private void deletePopupStart() {
 		frame = new JFrame("Delete Employee");
+
+		labelIdShowError = new JLabel();
+		labelIdShowError.setBounds(120, 10, 300, 30);
+		frame.add(labelIdShowError);
+
 		labelId = new JLabel("Enter Employee ID");
 		labelId.setBounds(20, 38, 200, 30);
 		frame.add(labelId);
@@ -57,20 +67,45 @@ public class DeletePopup {
 		frame.setLayout(null);
 		frame.setVisible(true);
 		frame.setResizable(false);
-		frame.setLocationRelativeTo(null);  
+		frame.setLocationRelativeTo(null);
 		frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 	}
 
-	//Database Operation call by this methods.
+	// Database Operation call by this methods.
 	private void deleteEmployeeById() {
 		String employeeId = textFieldId.getText();
-		boolean isDelete = DbRepository.delete(Long.parseLong(employeeId));
+		if (employeeId.isEmpty()) {
+			labelIdShowError.setFont(font);
+			labelIdShowError.setForeground(Color.RED);
+			labelIdShowError.setText("Please Enter the Id!!");
+		} else {
+			boolean isDelete = DbRepository.delete(Long.parseLong(employeeId));
 
-		if (isDelete == true) {
-			JOptionPane.showMessageDialog(null, "Data Deleted");
-			textFieldId.setText("");
-			frame.dispose();
+			if (isDelete == true) {
+				JOptionPane.showMessageDialog(null, "Data Deleted");
+				textFieldId.setText("");
+				frame.dispose();
+			}
 		}
+
+		// Textfield events Employee Name.
+		textFieldId.getDocument().addDocumentListener(new DocumentListener() {
+			@Override
+			public void changedUpdate(DocumentEvent arg0) {
+				labelIdShowError.setText(null);
+			}
+
+			@Override
+			public void insertUpdate(DocumentEvent arg0) {
+				labelIdShowError.setText(null);
+			}
+
+			@Override
+			public void removeUpdate(DocumentEvent arg0) {
+				labelIdShowError.setText(null);
+			}
+		});
+
 	}
 
 }
